Use confirmation system when deleting table
